refactor(_quality): group runtime / lint / validate modules into _quality/
Move four topically-related quality-gate modules into a new `_quality/`
subpackage to address the `scitex-dev ecosystem audit-all` PS-series
violation: 16 flat .py files at the package root, threshold 15. The
audit's own recommended layout names `_quality/` as the home for the
linter family — this PR is exactly that.

Moved
-----
- src/figrecipe/_validator.py                → src/figrecipe/_quality/_validator.py
- src/figrecipe/_axis_range_alignment.py     → src/figrecipe/_quality/_axis_range_alignment.py
- src/figrecipe/_axis_alignment_checker.py   → src/figrecipe/_quality/_axis_alignment_checker.py
- src/figrecipe/_linter_plugin.py            → src/figrecipe/_quality/_linter_plugin.py

Added
-----
- src/figrecipe/_quality/__init__.py         (package docstring; no re-exports —
                                              all consumers import the
                                              fully-qualified `_quality.<mod>`
                                              path directly)

Import updates
--------------
All import-path substitutions performed via `scitex-dev rename-symbols
--regex` (snapshot / dry-run / review / execute), three coordinated passes:

  1. Absolute  `figrecipe._<mod>`        → `figrecipe._quality._<mod>`
     (5 files / 6 matches — pyproject.toml entry-point, lazy-import inside
     _linter_plugin, docstring xref inside _axis_alignment_checker, two test
     files.)
  2. Relative  `from .._<mod>`           → `from .._quality._<mod>`
     (3 files / 4 matches — _api/_public.py, _api/_save.py, _api/_validate.py.)
  3. Relative  `from ._<mod>`            → `from ._quality._<mod>`
     (1 file / 1 match — utils.py.)

Total: 9 src files touched, 11 substitutions, 0 collisions.

Public surface
--------------
The figrecipe public API (figrecipe.fr.*, etc.) is unaffected — these are
all underscore-prefixed private modules. The figrecipe package
`__init__.py` does not re-export any of the moved modules, so no shim is
needed there. Tests use absolute `figrecipe._quality._<mod>` paths.

Flat-file count at src/figrecipe/ drops from 16 → 14 (well under the
threshold of 15). The four moved files now live under their canonical
topical home; the audit-all violation clears, which unblocks PR #138's
red CI gate (and every figrecipe PR after it).

Verify
------
  $ scitex-dev ecosystem audit-all figrecipe --severity error  # CLEAN
  $ pytest tests/figrecipe/_quality/                            # GREEN
  $ ls src/figrecipe/*.py | wc -l                               # 14 ≤ 15
