+++ src/figrecipe/_quality/_axis_range_alignment.py
