+++ src/figrecipe/_quality/_validator.py
